improve and refactor sql and backend with traze and observability

+++ transaction-back/src/main/java/com/example/transaction_back/exception/ResourceNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.transaction_back.exception;
 
 public class ResourceNotFoundException extends RuntimeException {
 
-    public ResourceNotFoundException(String resource, Integer id) {
+    public ResourceNotFoundException(String resource, Long id) {
         super(String.format("%s con id %d no encontrado", resource, id));
     }
 }
